Move plugin to cube_manipulation
